Add onModelChange callback prop to Layout
Add `onModelChange` callback prop to `Layout` to allow the parent component to know when the model changes.

Eventually I would like this to be a controlled component so that any changes to the model must be fed back in via the model prop in order to become active.  This would allow the state of `Layout` to live in an external store like Redux more cleanly.