fix: handle undefined references when page is unmounted
When FlexLayout is unmounted from the page, there is a race condition where some
FlexLayout internal code may attempt to access selfRef.current when it is
undefined (as the DOM has already been removed from the page).

The TypeScript code was using the ! operator to ignore the potential null values
in that field, which results in null-access errors when unmounting the page.

Various other locations where undefined reference errors were seen have also
been fixed to handle the respective variables being undefined.

https://github.com/caplin/FlexLayout/issues/352#issuecomment-1769322206

Fixes: #352

Signed-off-by: Christian Stewart <christian@aperture.us>