fix: make TabButton close button aria compliant as a button
Make the close button on TabButton a button instead of a div. It looks identical
and is aria compliant as a button, supporting screen readers and keyboard
navigation with tab to select and enter or space to click. Note that a event
handler for onCloseKeyDown was added to stop event propagation when pressing
enter or space, so that the button's normal onClick logic is called, and the
keyboard event does not propagate to the TabButton keyboard event handler.

Allows using the keyboard to select the tab close button by pressing the Tab key
to switch between elements on the page, and then pressing "enter" or space " "
to select the highlighted button.

Signed-off-by: Christian Stewart <christian@aperture.us>